requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.12.7
chardet==3.0.4
colorama==0.4.6
exceptiongroup==1.0.4
Faker==15.3.4
greenlet==2.0.1
idna==2.10
iniconfig==1.1.1
jsonpath==0.82
loguru
more-itertools==9.0.0
packaging==22.0
playwright==1.28.0
pluggy==0.13.1
py==1.11.0
pyee==9.0.4
PyMySQL==0.10.1
pytest==7.2.0
pytest-base-url==2.0.0
pytest-playwright==0.3.0
python-dateutil==2.8.2
python-slugify==7.0.0
PyYAML==6.0
requests==2.24.0
six==1.16.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.13
win32-setctime==1.1.0
xlrd==1.2.0
